+++ rbx_binary/src/core.rs
@@ -47,6 +47,20 @@ pub trait RbxReadExt: Read {
         Ok(f64::from_le_bytes(bytes))
     }
 
+    fn read_be_u32(&mut self) -> io::Result<u32> {
+        let mut bytes = [0; 4];
+        self.read_exact(&mut bytes)?;
+
+        Ok(u32::from_be_bytes(bytes))
+    }
+
+    fn read_be_i64(&mut self) -> io::Result<i64> {
+        let mut bytes = [0; 8];
+        self.read_exact(&mut bytes)?;
+
+        Ok(i64::from_be_bytes(bytes))
+    }
+
     fn read_u8(&mut self) -> io::Result<u8> {
         let mut buffer = [0u8];
         self.read_exact(&mut buffer)?;
@@ -82,6 +96,24 @@ pub trait RbxReadExt: Read {
 
     fn read_bool(&mut self) -> io::Result<bool> {
         Ok(self.read_u8()? != 0)
+    }
+
+    /// Fills `output` with blocks of `N` bytes from the buffer,
+    /// deinterleaving them in the process.
+    ///
+    /// This function allocates `N * output.len()` bytes before reading.
+    fn read_interleaved_bytes<const N: usize>(&mut self, output: &mut [[u8; N]]) -> io::Result<()> {
+        let len = output.len();
+        let mut buffer = vec![0; len * N];
+        self.read_exact(&mut buffer)?;
+
+        for (i, array) in output.iter_mut().enumerate() {
+            for (j, byte) in array.iter_mut().enumerate() {
+                *byte = buffer[i + len * j];
+            }
+        }
+
+        Ok(())
     }
 
     fn read_interleaved_i32_array(&mut self, output: &mut [i32]) -> io::Result<()> {
@@ -228,6 +260,23 @@ pub trait RbxWriteExt: Write {
 
     fn write_bool(&mut self, value: bool) -> io::Result<()> {
         self.write_u8(value as u8)
+    }
+
+    /// Takes `values` and writes it as a blob of data with each value
+    /// interleaved by `N` bytes.
+    ///
+    /// This function allocates `N * values.len()` bytes before writing.
+    fn write_interleaved_bytes<const N: usize>(&mut self, values: &[[u8; N]]) -> io::Result<()> {
+        let len = values.len();
+        let mut blob = vec![0; len * N];
+        for (x, bytes) in values.iter().enumerate() {
+            for (y, byte) in bytes.iter().enumerate() {
+                blob[y + x * N] = *byte;
+            }
+        }
+        self.write_all(&blob)?;
+
+        Ok(())
     }
 
     fn write_interleaved_i32_array<I>(&mut self, values: I) -> io::Result<()>