+++ rbx_binary/src/serializer/state.rs
@@ -11,7 +11,7 @@ use rbx_dom_weak::{
         Attributes, Axes, BinaryString, BrickColor, CFrame, Color3, Color3uint8, ColorSequence,
         ColorSequenceKeypoint, Content, Enum, Faces, Font, Matrix3, NumberRange, NumberSequence,
         NumberSequenceKeypoint, PhysicalProperties, Ray, Rect, Ref, SharedString, Tags, UDim,
-        UDim2, Variant, VariantType, Vector2, Vector3, Vector3int16,
+        UDim2, UniqueId, Variant, VariantType, Vector2, Vector3, Vector3int16,
     },
     Instance, WeakDom,
 };
@@ -1112,6 +1112,25 @@ impl<'dom, W: Write> SerializerState<'dom, W> {
                         chunk.write_u8(Type::Bool as u8)?;
                         chunk.write_all(bools.as_slice())?;
                     }
+                    Type::UniqueId => {
+                        let mut blobs = Vec::with_capacity(values.len());
+                        for (i, rbx_value) in values {
+                            if let Variant::UniqueId(value) = rbx_value.as_ref() {
+                                let mut blob = [0; 16];
+                                // This is maybe not the best solution to this
+                                // but we can always change it.
+                                blob[0..4].copy_from_slice(&value.index().to_be_bytes());
+                                blob[4..8].copy_from_slice(&value.time().to_be_bytes());
+                                blob[8..]
+                                    .copy_from_slice(&value.random().rotate_left(1).to_be_bytes());
+                                blobs.push(blob);
+                            } else {
+                                return type_mismatch(i, &rbx_value, "UniqueId");
+                            }
+                        }
+
+                        chunk.write_interleaved_bytes::<16>(&blobs)?;
+                    }
                 }
 
                 chunk.dump(&mut self.output)?;
@@ -1251,6 +1270,7 @@ impl<'dom, W: Write> SerializerState<'dom, W> {
             VariantType::Tags => Variant::Tags(Tags::new()),
             VariantType::Content => Variant::Content(Content::new()),
             VariantType::Attributes => Variant::Attributes(Attributes::new()),
+            VariantType::UniqueId => Variant::UniqueId(UniqueId::now().unwrap()),
             VariantType::Font => Variant::Font(Font::default()),
             _ => return None,
         })