+++ rbx_binary/src/tests/mod.rs
@@ -1,3 +1,4 @@
 mod models;
+mod places;
 mod serializer;
 mod util;