+++ rbx_binary/src/types.rs
@@ -39,6 +39,7 @@ pub enum Type {
     Int64 = 0x1B,
     SharedString = 0x1C,
     OptionalCFrame = 0x1E,
+    UniqueId = 0x1F,
     Font = 0x20,
 }
 
@@ -77,6 +78,7 @@ impl Type {
             VariantType::Int64 => Type::Int64,
             VariantType::SharedString => Type::SharedString,
             VariantType::OptionalCFrame => Type::OptionalCFrame,
+            VariantType::UniqueId => Type::UniqueId,
             VariantType::Font => Type::Font,
             _ => return None,
         })
@@ -113,6 +115,7 @@ impl Type {
             Type::Int64 => VariantType::Int64,
             Type::SharedString => VariantType::SharedString,
             Type::OptionalCFrame => VariantType::OptionalCFrame,
+            Type::UniqueId => VariantType::UniqueId,
             Type::Font => VariantType::Font,
         })
     }
@@ -152,6 +155,7 @@ impl TryFrom<u8> for Type {
             0x1B => Int64,
             0x1C => SharedString,
             0x1E => OptionalCFrame,
+            0x1F => UniqueId,
             0x20 => Font,
             _ => return Err(InvalidTypeError(value)),
         })