+++ rbx_xml/src/error.rs
@@ -92,6 +92,10 @@ pub(crate) enum DecodeErrorKind {
         actual_type: VariantType,
         message: String,
     },
+    InvalidPropertyData {
+        property_type: &'static str,
+        error: String,
+    },
 }
 
 impl fmt::Display for DecodeErrorKind {
@@ -138,6 +142,13 @@ impl fmt::Display for DecodeErrorKind {
                  When trying to convert, this error occured: {}",
                 class_name, property_name, expected_type, actual_type, message
             ),
+            InvalidPropertyData {
+                property_type,
+                error,
+            } => write!(
+                output,
+                "Could not decode property of type {property_type} because: {error}"
+            ),
         }
     }
 }
@@ -152,14 +163,7 @@ impl std::error::Error for DecodeErrorKind {
             ParseInt(err) => Some(err),
             DecodeBase64(err) => Some(err),
 
-            WrongDocVersion(_)
-            | UnexpectedEof
-            | UnexpectedXmlEvent(_)
-            | MissingAttribute(_)
-            | UnknownProperty { .. }
-            | InvalidContent(_)
-            | NameMustBeString(_)
-            | UnsupportedPropertyConversion { .. } => None,
+            _ => None,
         }
     }
 }