+++ rbx_xml/src/types/mod.rs
@@ -30,14 +30,15 @@ mod shared_string;
 mod strings;
 mod tags;
 mod udims;
+mod unique_id;
 mod vectors;
 
 use std::io::{Read, Write};
 
 use rbx_dom_weak::types::{
     Axes, BinaryString, CFrame, Color3, Color3uint8, ColorSequence, Content, Enum, Faces, Font,
-    NumberRange, NumberSequence, PhysicalProperties, Ray, Rect, Ref, UDim, UDim2, Variant, Vector2,
-    Vector2int16, Vector3, Vector3int16,
+    NumberRange, NumberSequence, PhysicalProperties, Ray, Rect, Ref, UDim, UDim2, UniqueId,
+    Variant, Vector2, Vector2int16, Vector3, Vector3int16,
 };
 
 use crate::{
@@ -146,6 +147,7 @@ declare_rbx_types! {
     String: String,
     UDim2: UDim2,
     UDim: UDim,
+    UniqueId: UniqueId,
     Vector2: Vector2,
     Vector2int16: Vector2int16,
     Vector3: Vector3,