+++ rbx_xml/CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ## 0.13.0 (2023-04-22)
 * Added support for `Font` values. ([#248])
+* Added support for `UniqueId` values. ([#256])
 
 [#248]: https://github.com/rojo-rbx/rbx-dom/pull/248
+[#256]: https://github.com/rojo-rbx/rbx-dom/pull/256
 
 ## 0.12.4 (2022-06-12)
 * Implemented serialization and deserialization for `Attributes`. ([#219])